Remove openssl engine flag from pure public key commands
Fixes an issue with OpenSSL 3.x which crashes on pkcs7_encrypt
with an engine flag set.
Removes the engine flag from all commands which neither involve a
private key nor random/parameter generation.
Removes the NEW_ALG option to use openssl with engine-provided
algorithms but as this is no longer supported in other parts of the
code this is (hopefully) only legacy from former GOST usage.

Commands affected:
 convert_cert, convert_pkcs10, convert_pkcs8, convert_pkey
 pkcs7_encrypt, pkcs7_decrypt